Isolate registration error
