+++ sdk/src/models/record-scanner/registrationResult.ts
@@ -1,24 +1,5 @@
 import type { RegistrationResponse } from "./registrationResponse.js";
-
-/**
- * Error thrown when a record scanner request fails (e.g. /register, /register/encrypted).
- * Includes HTTP status so callers can handle 422 vs 500 etc.
- */
-export class RecordScannerRequestError extends Error {
-    readonly status: number;
-
-    constructor(message: string, status: number) {
-        super(message);
-        this.name = "RecordScannerRequestError";
-        this.status = status;
-        Object.setPrototypeOf(this, RecordScannerRequestError.prototype);
-    }
-}
-
-/** Error payload for registration failure result (/register and /register/encrypted). */
-export interface RegistrationErrorBody {
-    message: string;
-}
+import type { RecordScannerErrorBody } from "./error.js";
 
 /** Success variant of registration result. */
 export interface RegisterSuccess {
@@ -30,7 +11,7 @@ export interface RegisterSuccess {
 export interface RegisterFailure {
     ok: false;
     status: number;
-    error: RegistrationErrorBody;
+    error: RecordScannerErrorBody;
 }
 
 /** Result of register() and registerEncrypted(); never throws on HTTP error. */
